% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_zoom.R
\name{nhm_map_reset}
\alias{nhm_map_reset}
\title{Reset a plotly geo map to the global view}
\usage{
nhm_map_reset(session, outputId, duration = 1200)
}
\arguments{
\item{session}{The Shiny session object.}

\item{outputId}{The output ID of the \code{plotlyOutput} to animate.}

\item{duration}{Animation duration in milliseconds. Default 1200.}
}
\value{
Called for its side effect.
}
\description{
Flies back to the default global view (lat 0, lon 0, scale 1).
}
