% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\name{scale_colour_nhm}
\alias{scale_colour_nhm}
\alias{scale_color_nhm}
\title{NHM discrete colour scale}
\usage{
scale_colour_nhm(...)

scale_color_nhm(...)
}
\arguments{
\item{...}{Arguments passed to \code{ggplot2::discrete_scale}.}
}
\value{
A ggplot2 colour scale.
}
\description{
A colour scale using the NHM palette for discrete aesthetics.
}
