% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\name{scale_fill_nhm}
\alias{scale_fill_nhm}
\title{NHM discrete fill scale}
\usage{
scale_fill_nhm(...)
}
\arguments{
\item{...}{Arguments passed to \code{ggplot2::discrete_scale}.}
}
\value{
A ggplot2 fill scale.
}
\description{
A fill scale using the NHM palette for discrete aesthetics.
}
